requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
sqlparse==0.4.3
coverage==6.5.0
django-widget-tweaks==1.4.12
libgravatar==1.0.3
Faker==17.6.0
python-dateutil==2.8.2
six==1.16.0
scikit-surprise==1.1.3